chore(cli): move command usage helper into core
Relocate resolveCommandUsage from src/cli/shared/command-usage.js to src/cli/core/command-usage.js so the Commander-specific helper lives alongside the rest of the command manager utilities.
